requirements.txt
python-telegram-bot==20.8
psycopg2-binary==2.9.9
requests==2.32.3
